+++ src/helpers/index.tsx
@@ -16,7 +16,7 @@ import { ReactComponent as SOhmImg } from "../assets/tokens/token_sOHM.svg";
 import { addresses, BLOCK_RATE_SECONDS, EPOCH_INTERVAL, NetworkId } from "../constants";
 import { PairContract, RedeemHelper } from "../typechain";
 import { ohm_dai, ohm_daiOld, ohm_weth } from "./AllBonds";
-import { Environment } from "./environment/environment";
+import { Environment } from "./environment/Environment";
 import { Providers } from "./providers/providers";
 
 /**
