+++ src/hooks/useSegmentAnalytics.ts
@@ -5,7 +5,7 @@ import { useLocation } from "react-router-dom";
 import { useWeb3Context } from "src/hooks/web3Context";
 import { v4 as uuidv4 } from "uuid";
 
-import { Environment } from "../helpers/environment/environment";
+import { Environment } from "../helpers/environment/Environment";
 import { getParameterByName } from "../helpers/QueryParameterHelper";
 
 const SEGMENT_API_KEY = Environment.getSegmentApiKey();
