+++ src/views/Give/RecipientModal.tsx
@@ -13,7 +13,7 @@ import { useLocation } from "react-router-dom";
 import { Project } from "src/components/GiveProject/project.type";
 import { NetworkId } from "src/constants";
 import { shorten } from "src/helpers";
-import { Environment } from "src/helpers/environment/environment";
+import { Environment } from "src/helpers/environment/Environment";
 import { useWeb3Context } from "src/hooks/web3Context";
 import {
   changeApproval,
